# projects/bravo/docs/openapi_api_contracts.yaml
openapi: 3.1.0
info:
  title: Bravo API
  description: |
    Bravo - "Championing Our Data" - Jira hygiene nudge system.

    Monitors Jira tickets for watched assignees, detects incomplete documentation,
    and nudges engineers via Slack to add updates.
  version: 0.1.0
  contact:
    name: Campaign Operations Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: http://bravo-api:8000
    description: Docker internal
...
paths:
...
  /nudge:
    get:
      tags: [Nudge]
      summary: List nudge events
      operationId: listNudges
      parameters:
        - name: ticket_key
          in: query
          schema:
            type: string
        - name: assignee
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/NudgeStatus'
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of nudge events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NudgeListResponse'
  /nudge/{nudge_id}:
    get:
      tags: [Nudge]
      summary: Get nudge event details
      operationId: getNudge
      parameters:
        - name: nudge_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Nudge event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NudgeResponse'

  # ============ ASSIGNEES ============
...
components:
  schemas:
...
    NudgeStatus:
      type: string
      enum: [SENT, ACKNOWLEDGED, SNOOZED, RESPONDED, POSTED, CANCELLED]
    NudgeResponse:
      type: object
      properties:
        nudge_id:
          type: string
          format: uuid
        ticket_key:
          type: string
        assignee_jira_id:
          type: string
        status:
          $ref: '#/components/schemas/NudgeStatus'
        trigger_reason:
          type: string
          description: Which gate failed or LLM score below threshold
        slack_channel:
          type: string
        slack_ts:
          type: string
        message_content:
          type: string
        created_at:
          type: string
          format: date-time
        responded_at:
          type: string
          format: date-time
          nullable: true
        jira_comment_posted:
          type: string
          nullable: true
          description: The comment text posted to Jira (if any)
    NudgeListResponse:
      type: object
      properties:
        nudges:
          type: array
          items:
            $ref: '#/components/schemas/NudgeResponse'
        total:
          type: integer

    # ============ ASSIGNEES ============
...